backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
requests>=2.28.0
python-decouple>=3.8
geopy>=2.3.0
python-dateutil>=2.8.0